Completed update test for insurance resource
Also moved test back outside container for now. It is somewhat complex
to use a mock request to run the request on the test thread so we can
run it transactional so for now we run 'real' requests.

If we're running real requests we might as well run them from outside so
we can more easily debug the test.